# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GreenVolt Solar Panel Service Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all service jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed]
        - name: staffId
          in: query
          schema:
            type: integer
        - name: customerId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of service jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListResponse"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    put:
      operationId: updateService
      tags: [services]
      summary: Update service job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceInput"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete service job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, success, failed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create/record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    put:
      operationId: updatePayment
      tags: [payments]
      summary: Update payment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentInput"
      responses:
        "200":
          description: Payment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        solarCapacity:
          type: number
          nullable: true
        installationDate:
          type: string
          nullable: true
        installationDetails:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - phone
        - address
        - createdAt
        - updatedAt
...
    Staff:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        workArea:
          type: string
          nullable: true
        availability:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - phone
        - role
        - isActive
        - createdAt
        - updatedAt
...
    Service:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        staffId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed]
        scheduledDate:
          type: string
        notes:
          type: string
          nullable: true
        serviceType:
          type: string
          nullable: true
        preServiceImage:
          type: string
          nullable: true
        postServiceImage:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        customer:
          $ref: "#/components/schemas/Customer"
          nullable: true
        staff:
          $ref: "#/components/schemas/Staff"
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerId
        - status
        - scheduledDate
        - createdAt
        - updatedAt
    CreateServiceInput:
      type: object
      properties:
        customerId:
          type: integer
        staffId:
          type: integer
        scheduledDate:
          type: string
        notes:
          type: string
        serviceType:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed]
      required:
        - customerId
        - scheduledDate
    UpdateServiceInput:
      type: object
      properties:
        staffId:
          type: integer
        status:
          type: string
          enum: [pending, in_progress, completed]
        scheduledDate:
          type: string
        notes:
          type: string
        serviceType:
          type: string
        preServiceImage:
          type: string
        postServiceImage:
          type: string
        remarks:
          type: string
        completedAt:
          type: string
    ServiceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        subscriptionId:
          type: integer
          nullable: true
        amount:
          type: number
        status:
          type: string
          enum: [pending, success, failed]
        paymentMethod:
          type: string
          nullable: true
        transactionId:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        customer:
          $ref: "#/components/schemas/Customer"
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerId
        - amount
        - status
        - createdAt
        - updatedAt
    CreatePaymentInput:
      type: object
      properties:
        customerId:
          type: integer
        subscriptionId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, success, failed]
        paymentMethod:
          type: string
        transactionId:
          type: string
        description:
          type: string
      required:
        - customerId
        - amount
        - status
    UpdatePaymentInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, success, failed]
        transactionId:
          type: string
        paymentMethod:
          type: string
    PaymentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT